requirements.txt
psycopg2
Django==4.1
djangorestframework
requests
